Pipeline and services refactor 1: resume parsing
